Alex li bug fix

See merge request tevar/jewelrycg!40

+++ database/migrations/2022_12_05_121906_create_product_measurements_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('product_measurements', function (Blueprint $table) {
             $table->id();
-            $table->decimal('name', 10, 2)->default(0);
+            $table->integer('name')->default(0);
             $table->string('units');
             $table->timestamps();
         });
